Tried to use a dropdown list in a crudevents.html

+++ src/main/java/accountingApp/controller/EventsController.java
@@ -27,39 +27,59 @@ public class EventsController {
     @GetMapping("/events")
     public String getEvents(Model model) {
         List<Events> eventsList = eventsService.findAllEvents();
+        List<Employee> employeeList = employeeService.getListEmployee();
+        List<ITStaff> itStaffList = itStaffService.getAllItStaff();
+        List<WorkArea> workAreaList = workAreaService.findAllWorkArea();
+        List<Devices> devicesList = devicesService.findAllDevices();
         model.addAttribute("eventsList", eventsList);
+        model.addAttribute("employeeList", employeeList);
+        model.addAttribute("itStaffList", itStaffList);
+        model.addAttribute("workAreaList", workAreaList);
+        model.addAttribute("devicesList", devicesList);
         return "events";
     }
 
     @PostMapping("/addevent")
     public String addNewEvent(@RequestParam String date,
-                              @RequestParam String device,
-                              @RequestParam String employeeid,
-                              @RequestParam String itstaffid,
-                              @RequestParam String workarea,
+                              @RequestParam Devices device,
+                              @RequestParam Employee employee,
+                              @RequestParam ITStaff itstaff,
+                              @RequestParam WorkArea workarea,
                               @RequestParam String comment,
                               Model model) {
+
+        if (date == null
+                || device == null
+                || employee == null
+                || itstaff == null
+                || workarea == null
+                || comment == null
+        ) {
+            System.out.println("*** EventsController.addEvent:  Attribute has a null value! ***");
+            return this.getEvents(model);
+        }
+
         String dateWithoutSpaces = date.trim();
-        String deviceWithoutSpaces = device.trim();
-        String employeeidWithoutSpaces = employeeid.trim();
-        String itstaffidWithoutSpaces = itstaffid.trim();
-        String workareaWithoutSpaces = workarea.trim();
+        String deviceWithoutSpaces = device.getName();
+        String employeeIdWithoutSpaces = employee.getFio();
+        String itstaffIdWithoutSpaces = itstaff.getName();
+        String workareaWithoutSpaces = workarea.getName();
         String commentWithoutSpaces = comment.trim();
         try {
             int dateCheck = Integer.parseInt(dateWithoutSpaces);
             if (dateCheck <= 0 || date.length() < 6) {
-                System.out.println("*** WRONG DATE ***");
-            } else {
-                List<Devices> devicesList = devicesService.getDevicesByName(deviceWithoutSpaces);
-                List<Employee> employeeList = employeeService.findEmployeeByFio(employeeidWithoutSpaces);
-                List<ITStaff> itStaffList = itStaffService.getITStaffByName(itstaffidWithoutSpaces);
-                List<WorkArea> workAreaList = workAreaService.getWorkAreaById(
-                        Integer.parseInt(workareaWithoutSpaces));
+                System.out.println("*** EventsController.addNewEvent:  WRONG DATE FORMAT***");
+            } else if (!dateWithoutSpaces.equals("")
+                    && !deviceWithoutSpaces.equals("")
+                    && !employeeIdWithoutSpaces.equals("")
+                    && !workareaWithoutSpaces.equals("")
+                    && !itstaffIdWithoutSpaces.equals("")) {
+
                 Events events = new Events(date,
-                        devicesList.get(0),
-                        employeeList.get(0),
-                        itStaffList.get(0),
-                        workAreaList.get(0),
+                        device,
+                        employee,
+                        itstaff,
+                        workarea,
                         commentWithoutSpaces);
                 eventsService.addNewEvent(events);
                 List<Events> eventsList = eventsService.findAllEvents();
@@ -67,7 +87,8 @@ public class EventsController {
             }
             return this.getEvents(model);
         } catch (Exception e) {
-            System.out.println("*** WRONG DATE FORMAT ***");
+            System.out.println("*** EventsController.addNewEvent:  WRONG DB VALUES*** "
+                    + e.getMessage());
             return this.getEvents(model);
         }
     }
@@ -75,10 +96,16 @@ public class EventsController {
     @PostMapping("/deleteevent")
     public String deleteEvent(@RequestParam String id,
                               Model model) {
+        if (id == null
+        ) {
+            System.out.println("*** EventsController.deleteEvent:  Attribute ID has a null value! ***");
+            return this.getEvents(model);
+        }
+
         try {
             int idCheck = Integer.parseInt(id);
             if (idCheck <= 0) {
-                System.out.println("*** SUB ZERO ID***");
+                System.out.println("*** EventsController.addNewEvent: ID is SUBZERO***");
             } else {
                 eventsService.deleteEventsById(idCheck);
                 List<Events> eventsList = eventsService.findAllEvents();
@@ -86,7 +113,8 @@ public class EventsController {
             }
             return this.getEvents(model);
         } catch (Exception e) {
-            System.out.println("*** WRONG ID FORMAT***");
+            System.out.println("*** EventsController.deleteEvent:  WRONG DB VALUES*** "
+                    + e.getMessage());
             return this.getEvents(model);
         }
 
@@ -95,40 +123,51 @@ public class EventsController {
     @PostMapping("/updateevent")
     public String updateEvent(@RequestParam String id,
                               @RequestParam String date,
-                              @RequestParam String device,
-                              @RequestParam String employeeid,
-                              @RequestParam String itstaffid,
-                              @RequestParam String workarea,
+                              @RequestParam Devices device,
+                              @RequestParam Employee employee,
+                              @RequestParam ITStaff itstaff,
+                              @RequestParam WorkArea workarea,
                               @RequestParam String comment,
                               Model model) {
 
+        if (id == null
+                || date == null
+                || device == null
+                || employee == null
+                || itstaff == null
+                || workarea == null
+                || comment == null
+        ) {
+            System.out.println("*** EventsController.updateEvent:  Attribute has a null value! ***");
+            return this.getEvents(model);
+        }
+
         String idWithoutSpaces = id.trim();
         String dateWithoutSpaces = date.trim();
-        String deviceWithoutSpaces = device.trim();
-        String employeeIdWithoutSpaces = employeeid.trim();
-        String itstaffIdWithoutSpaces = itstaffid.trim();
-        String workareaWithoutSpaces = workarea.trim();
+        String deviceWithoutSpaces = device.getName();
+        String employeeIdWithoutSpaces = employee.getFio();
+        String itstaffIdWithoutSpaces = itstaff.getName();
+        String workareaWithoutSpaces = workarea.getName();
         String commentWithoutSpaces = comment.trim();
+
         try {
             int idCheck = Integer.parseInt(idWithoutSpaces);
-            if (idCheck <= 0) {
-                System.out.println("*** SUB ZERO ID***");
+            int dateCheck = Integer.parseInt(dateWithoutSpaces);
+            if (dateCheck <= 0 || date.length() < 6 || idCheck <= 0) {
+                System.out.println("*** EventsController.updateEvent:  WRONG DATE FORMAT or ID FORMAT***");
                 return this.getEvents(model);
-            } else if (!dateWithoutSpaces.equals("") &&
-                    !deviceWithoutSpaces.equals("") &&
-                    !employeeIdWithoutSpaces.equals("") &&
-                    !itstaffIdWithoutSpaces.equals("")) {
-                List<Devices> devicesList = devicesService.getDevicesByName(deviceWithoutSpaces);
-                List<Employee> employeeList = employeeService.findEmployeeByFio(employeeIdWithoutSpaces);
-                List<ITStaff> itStaffList = itStaffService.getITStaffByName(itstaffIdWithoutSpaces);
-                List<WorkArea> workAreaList = workAreaService
-                        .getWorkAreaById(Integer.parseInt(workareaWithoutSpaces));
+            } else if (!dateWithoutSpaces.equals("")
+                    && !deviceWithoutSpaces.equals("")
+                    && !employeeIdWithoutSpaces.equals("")
+                    && !workareaWithoutSpaces.equals("")
+                    && !itstaffIdWithoutSpaces.equals("")) {
+
                 Events events = new Events(idCheck,
                         dateWithoutSpaces,
-                        devicesList.get(0),
-                        employeeList.get(0),
-                        itStaffList.get(0),
-                        workAreaList.get(0),
+                        device,
+                        employee,
+                        itstaff,
+                        workarea,
                         commentWithoutSpaces);
                 eventsService.updateEvent(events);
                 List<Events> eventsList = eventsService.findAllEvents();
@@ -137,21 +176,27 @@ public class EventsController {
             return this.getEvents(model);
 
         } catch (Exception e) {
-            System.out.println("*** WRONG ID FORMAT***");
-            this.getEvents(model);
-            return this.getEvents(model);
+            System.out.println("*** EventsController.updateEvent:  WRONG DB VALUES*** "
+                    + e.getMessage());
+            getEvents(model);
+            return getEvents(model);
         }
     }
 
     @PostMapping("/findeventbyid")
     public String findEventsById(@RequestParam String id,
                                  Model model) {
+        if (id == null
+        ) {
+            System.out.println("*** EventsController.findEventsById:  Attribute ID has a null value! ***");
+            return this.getEvents(model);
+        }
 
         String idWithoutSpaces = id.trim();
         try {
             int idCheck = Integer.parseInt(idWithoutSpaces);
             if (idCheck <= 0) {
-                System.out.println("*** SUB ZERO ID***");
+                System.out.println("*** EventsController.fidEventById: ID is SUBZERO***");
                 this.getEvents(model);
             } else {
                 List<Events> eventsList;
@@ -160,7 +205,8 @@ public class EventsController {
             }
             return "events";
         } catch (Exception e) {
-            System.out.println("*** WRONG ID FORMAT***");
+            System.out.println("*** EventsController.fidEventById:  WRONG DB VALUES*** "
+                    + e.getMessage());
             return this.getEvents(model);
         }
